refactor: front-end utils
