+++ frontend/src/components/UserSettings/ChangePassword.tsx
@@ -14,7 +14,7 @@ import { type SubmitHandler, useForm } from "react-hook-form"
 
 import { type ApiError, type UpdatePassword, UsersService } from "../../client"
 import useCustomToast from "../../hooks/useCustomToast"
-import { confirmPasswordRules, handleError, passwordRules } from "../../utils"
+import { confirmPasswordRules, handleError, passwordRules } from "../../utils/core"
 
 interface UpdatePasswordForm extends UpdatePassword {
   confirm_password: string
